Refactor: Class files
- Each class has a own swift file
- Remove break lines
