+++ CardGameApp/CardGameApp/Views/PokerGame/PlayerStackView.swift
@@ -1,0 +1,35 @@
+//
+//  PlayerStackView.swift
+//  CardGameApp
+//
+//  Created by Cory Kim on 2020/02/12.
+//  Copyright © 2020 corykim0829. All rights reserved.
+//
+
+import UIKit
+
+class PlayerStackView: UIStackView {
+    private var displayName: String = ""
+    let nameLabel = UILabel()
+    
+    init(displayName: String) {
+        self.displayName = displayName
+        super.init(frame: .zero)
+        commonInit()
+    }
+    
+    required init(coder: NSCoder) {
+        super.init(coder: coder)
+        commonInit()
+    }
+    
+    private func commonInit() {
+        axis = .vertical
+        spacing = 4
+        translatesAutoresizingMaskIntoConstraints = false
+        
+        nameLabel.textColor = .white
+        nameLabel.text = displayName
+        addArrangedSubview(nameLabel)
+    }
+}
