+++ CardGameApp/CardGameApp/Views/PokerGame/PokerGameStackView.swift
@@ -1,0 +1,28 @@
+//
+//  PokerGameStackView.swift
+//  CardGameApp
+//
+//  Created by Cory Kim on 2020/02/12.
+//  Copyright © 2020 corykim0829. All rights reserved.
+//
+
+import UIKit
+
+class PokerGameStackView: UIStackView {
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        commonInit()
+    }
+    
+    required init(coder: NSCoder) {
+        super.init(coder: coder)
+        commonInit()
+    }
+    
+    private func commonInit() {
+        axis = .vertical
+        spacing = 16
+        translatesAutoresizingMaskIntoConstraints = false
+        alignment = .top
+    }
+}
