+++ CardGameApp/CardGameApp/Views/TitledSegmentedControl.swift
@@ -1,8 +1,8 @@
 //
-//  SegmentedControls.swift
+//  TitledSegmentedControl.swift
 //  CardGameApp
 //
-//  Created by Cory Kim on 2020/02/11.
+//  Created by Cory Kim on 2020/02/12.
 //  Copyright © 2020 corykim0829. All rights reserved.
 //
 
@@ -32,24 +32,5 @@ class TitledSegmentedControl: UISegmentedControl {
         items?.enumerated().forEach { (index, item) in
             insertSegment(withTitle: item, at: index, animated: false)
         }
-    }
-}
-
-class SegmentedControlStackView: UIStackView {
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        commonInit()
-    }
-    
-    required init(coder: NSCoder) {
-        super.init(coder: coder)
-        commonInit()
-    }
-    
-    private func commonInit() {
-        axis = .vertical
-        spacing = 8
-        distribution = .fillEqually
-        translatesAutoresizingMaskIntoConstraints = false
     }
 }
